+++ webpack.packages.json
@@ -35,7 +35,6 @@
 
   "redux-devtools": "^3.0.1",
   "redux-devtools-log-monitor": "^1.0.2",
-  "redux-devtools-dock-monitor": "^1.0.1",
-  "history": "^2.0.0-rc2"
+  "redux-devtools-dock-monitor": "^1.0.1"
 
 }
